+++ driver-core/src/main/java/com/datastax/driver/core/DefaultResultSetFuture.java
@@ -71,60 +71,38 @@ class DefaultResultSetFuture extends AbstractFuture<ResultSet> implements Result
                             ResultSet rs = ArrayBackedResultSet.fromMessage(rm, session, protocolVersion, info, statement);
                             switch (scc.change) {
                                 case CREATED:
-                                    switch (scc.target) {
-                                        case KEYSPACE:
-                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, null);
-                                            break;
-                                        case TABLE:
-                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, scc.name, null);
-                                            break;
-                                        case TYPE:
-                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, scc.name);
-                                            break;
-                                    }
+                                case UPDATED:
+                                    session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.targetType, scc.targetKeyspace, scc.targetName);
                                     break;
                                 case DROPPED:
                                     KeyspaceMetadata keyspace;
-                                    switch (scc.target) {
+                                    switch (scc.targetType) {
                                         case KEYSPACE:
                                             // If that the one keyspace we are logged in, reset to null (it shouldn't really happen but ...)
                                             // Note: Actually, Cassandra doesn't do that so we don't either as this could confuse prepared statements.
                                             // We'll add it back if CASSANDRA-5358 changes that behavior
                                             //if (scc.keyspace.equals(session.poolsState.keyspace))
                                             //    session.poolsState.setKeyspace(null);
-                                            session.cluster.manager.metadata.removeKeyspace(scc.keyspace);
+                                            session.cluster.manager.metadata.removeKeyspace(scc.targetKeyspace);
                                             break;
                                         case TABLE:
-                                            keyspace = session.cluster.manager.metadata.getKeyspaceInternal(scc.keyspace);
+                                            keyspace = session.cluster.manager.metadata.getKeyspaceInternal(scc.targetKeyspace);
                                             if (keyspace == null)
                                                 logger.warn("Received a DROPPED notification for table {}.{}, but this keyspace is unknown in our metadata",
-                                                    scc.keyspace, scc.name);
+                                                    scc.targetKeyspace, scc.targetName);
                                             else
-                                                keyspace.removeTable(scc.name);
+                                                keyspace.removeTable(scc.targetName);
                                             break;
                                         case TYPE:
-                                            keyspace = session.cluster.manager.metadata.getKeyspaceInternal(scc.keyspace);
+                                            keyspace = session.cluster.manager.metadata.getKeyspaceInternal(scc.targetKeyspace);
                                             if (keyspace == null)
                                                 logger.warn("Received a DROPPED notification for UDT {}.{}, but this keyspace is unknown in our metadata",
-                                                    scc.keyspace, scc.name);
+                                                    scc.targetKeyspace, scc.targetName);
                                             else
-                                                keyspace.removeUserType(scc.name);
+                                                keyspace.removeUserType(scc.targetName);
                                             break;
                                     }
                                     session.cluster.manager.waitForSchemaAgreementAndSignal(connection, this, rs);
-                                    break;
-                                case UPDATED:
-                                    switch (scc.target) {
-                                        case KEYSPACE:
-                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, null);
-                                            break;
-                                        case TABLE:
-                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, scc.name, null);
-                                            break;
-                                        case TYPE:
-                                            session.cluster.manager.refreshSchemaAndSignal(connection, this, rs, scc.keyspace, null, scc.name);
-                                            break;
-                                    }
                                     break;
                                 default:
                                     logger.info("Ignoring unknown schema change result");