+++ driver-core/CHANGELOG.rst
@@ -1,6 +1,20 @@
 CHANGELOG
 =========
 
+2.2.0-rc1:
+----------
+
+Merged from 2.1 branch:
+
+- [improvement] Unify "Target" enum for schema elements (JAVA-782)
+
+
+2.1.7:
+------
+
+- [improvement] Unify "Target" enum for schema elements (JAVA-782)
+
+
 2.1.6:
 ------